Compiler warnings

+++ src/network/messages/otp_module_advertisement_message.cpp
@@ -81,18 +81,18 @@ Message::Message(
 bool Message::isValid()
 {
     size_t lengthCheck = toByteArray().length();
-    if (lengthCheck != otpLayer->getPDULength() + OTPLayer::LENGTHOFFSET)
+    if (lengthCheck != static_cast<size_t>(otpLayer->getPDULength() + OTPLayer::LENGTHOFFSET))
         return false;
     if (!otpLayer->isValid()) return false;
 
     lengthCheck -= otpLayer->toPDUByteArray().length();
-    if (lengthCheck != advertisementLayer->getPDULength() + OTPAdvertisementLayer::LENGTHOFFSET)
+    if (lengthCheck != static_cast<size_t>(advertisementLayer->getPDULength() + OTPAdvertisementLayer::LENGTHOFFSET))
         return false;
     if (advertisementLayer->getVector() != PDU::VECTOR_OTP_ADVERTISEMENT_MODULE) return false;
     if (!advertisementLayer->isValid()) return false;
 
     lengthCheck -= advertisementLayer->toPDUByteArray().length();
-    if (lengthCheck != moduleAdvertisementLayer->getPDULength() + OTPModuleAdvertisementLayer::LENGTHOFFSET)
+    if (lengthCheck != static_cast<size_t>(moduleAdvertisementLayer->getPDULength() + OTPModuleAdvertisementLayer::LENGTHOFFSET))
         return false;
     if (!moduleAdvertisementLayer->isValid()) return false;
     if (!RANGES::MESSAGE_SIZE.isValid(toByteArray().size())) return false;
